Developed new crowded center 9x9 and 11x11, crowded periphery 11x11, restructured data output
