--- Crowded Center/Data/Anna_01-17_Crowded Center Flipped V2.csv
@@ -1,15 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-1,0.7000000000000002,15
-1,0.7000000000000002,25
-1,1.4,35
-1,0.4,10
-1,2.0999999999999996,40
-1,0.7000000000000002,20
-1,1.0,30
-3,0.8,30
-3,0.40000000000000013,15
-3,0.7000000000000002,20
-3,0.20000000000000004,10
-3,0.8,25
-3,1.4999999999999998,40
-3,1.7,35
