--- Crowded Center/Data/Anna_01-17_Crowded Center Flipped.csv
@@ -1,15 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-3,1.0,30
-3,0.4,10
-3,0.9999999999999998,35
-3,1.2,25
-3,0.5,20
-3,2.1,40
-3,0.40000000000000024,15
-1,0.4,10
-1,0.5000000000000002,25
-1,0.8,30
-1,1.3,35
-1,0.3000000000000002,15
-1,0.9999999999999998,40
-1,0.7000000000000002,20
