--- Crowded Center/Data/Jane_01-17_Crowded Center Flipped.csv
@@ -1,25 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-4,0.5000000000000002,15
-4,0.7999999999999998,30
-4,1.1,25
-4,0.5000000000000002,20
-4,0.30000000000000004,10
-1,0.29999999999999993,20
-1,0.30000000000000004,25
-1,0.20000000000000004,15
-1,0.39999999999999997,35
-1,0.30000000000000004,10
-1,0.5,40
-1,0.20000000000000004,30
-2,0.6,25
-2,0.20000000000000004,10
-2,0.30000000000000016,20
-2,0.4,15
-2,0.3999999999999999,30
-3,0.40000000000000013,25
-3,0.40000000000000024,35
-3,0.4,30
-3,0.30000000000000004,20
-3,0.30000000000000004,10
-3,0.5,40
-3,0.30000000000000004,15
