--- Crowded Center/Data/Thida3by3_01-17_Crowded Center Flipped V2.csv
@@ -1,15 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-3,0.4,10
-3,0.5000000000000002,30
-3,0.30000000000000027,20
-3,0.40000000000000013,25
-3,1.4,40
-3,1.2,35
-3,0.40000000000000013,15
-1,0.40000000000000013,10
-1,0.8,25
-1,1.3999999999999997,40
-1,0.7000000000000002,15
-1,0.7000000000000002,20
-1,1.1999999999999997,35
-1,1.2,30
