--- Crowded Center/Data/ThidaREAL_01-17_Crowded Center Flipped.csv
@@ -1,15 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-3,0.2000000000000002,25
-3,0.2000000000000002,30
-3,0.7000000000000002,35
-3,0.3000000000000001,15
-3,0.30000000000000004,10
-3,0.40000000000000013,40
-3,0.30000000000000004,20
-1,0.4,15
-1,0.5000000000000002,30
-1,0.9000000000000001,25
-1,1.4,40
-1,0.30000000000000004,10
-1,0.5,35
-1,0.20000000000000004,20
