--- Crowded Center/Data/Thida_01-17_Crowded Center Flipped.csv
@@ -1,12 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-3,0.40000000000000024,25
-3,0.4,15
-3,0.30000000000000004,10
-3,0.30000000000000027,30
-3,0.20000000000000043,40
-3,0.40000000000000013,20
-3,0.8999999999999999,35
-1,0.20000000000000026,35
-1,0.7000000000000002,15
-1,0.30000000000000027,20
-1,0.3000000000000002,25
