--- Crowded Periphery 5x5/Data/placeholder.txt
@@ -1,6 +1,0 @@
-{\rtf1\ansi\ansicpg1252\cocoartf1671
-{\fonttbl}
-{\colortbl;\red255\green255\blue255;}
-{\*\expandedcolortbl;;}
-\margl1440\margr1440\vieww10800\viewh8400\viewkind0
-}
