--- Isolated Character/Data/AnnaDSremapkeys_01-13_Isolated Character Flipped.csv
@@ -1,9 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-1,1.9,40
-1,0.7000000000000002,30
-1,0.4,10
-1,0.40000000000000013,20
-3,0.7000000000000002,30
-3,1.4,40
-3,0.4,10
-3,0.5000000000000002,20
