--- Isolated Character/Data/Anna_01-13_Isolated Character Flipped.csv
@@ -1,9 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-1,1.0,30
-1,0.7000000000000002,20
-1,1.7,40
-1,0.4,10
-3,1.5999999999999999,40
-3,0.4,10
-3,0.40000000000000024,20
-3,0.8999999999999999,30
