--- Isolated Character/Data/Anna_01-20_Isolated Character.csv
@@ -1,19 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-1,0.20000000000000004,0
-1,0.5000000000000002,20
-1,0.30000000000000004,5
-1,1.0,25
-1,0.4,10
-1,0.3000000000000002,15
-1,1.7,35
-1,1.5,40
-1,1.0,30
-3,0.30000000000000004,5
-3,0.40000000000000013,10
-3,0.20000000000000004,0
-3,1.2,30
-3,2.1000000000000005,40
-3,0.8,25
-3,1.2,35
-3,0.7000000000000002,15
-3,1.0999999999999999,20
