--- Isolated Character/Data/Ethan_01-14_Isolated Character Flipped.csv
@@ -1,11 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-3,0.1,20
-3,0.5,30
-3,0.2,0
-3,0.5,10
-3,1.7,40
-1,0.20000000000000004,10
-1,0.3000000000000002,20
-1,0.8,30
-1,1.0,40
-1,0.2,0
