--- Isolated Character/Data/Jane_01-15_Isolated Character Flipped.csv
@@ -1,17 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-2,0.4,10
-2,0.6000000000000001,30
-2,0.40000000000000013,20
-3,0.7000000000000002,20
-3,0.20000000000000004,0
-3,1.5999999999999999,40
-3,0.7000000000000002,30
-3,0.4,10
-1,0.5,20
-1,0.30000000000000004,10
-1,0.4,30
-1,0.9999999999999998,40
-1,0.20000000000000004,0
-4,0.4,10
-4,1.7,30
-4,0.4000000000000001,20
