--- Isolated Character/Data/thida_01-13_Isolated Character Flipped.csv
@@ -1,11 +1,0 @@
-Direction,Letter Height (degrees),Eccentricity (degrees)
-3,1.2,40
-3,1.0,30
-3,0.40000000000000013,10
-3,0.4,20
-3,0.20000000000000004,0
-1,1.2,40
-1,0.5,20
-1,0.7000000000000002,30
-1,0.30000000000000004,0
-1,0.3999999999999999,10
