--- T1/Data/Anna_11-5_T1.csv
@@ -1,181 +1,0 @@
-Direction, Characters Read, Eccentricity (degrees), Letter Height (degrees)
-2,2,0.69,0.23
-2,1,0.34,0.23
-2,2,0.69,0.23
-2,3,2.06,0.5
-2,5,3.43,0.5
-2,4,2.75,0.5
-2,3,4.19,1
-2,3,4.19,1
-2,4,5.57,1
-2,3,6.23,1.5
-2,4,8.28,1.5
-2,4,8.28,1.5
-2,3,8.19,2
-2,3,8.19,2
-2,4,10.87,2
-2,3,10.27,2.5
-2,0,0,2.5
-2,5,16.8,2.5
-2,4,16.4,3
-2,4,16.4,3
-2,4,16.4,3
-2,2,9.76,3.5
-2,2,9.76,3.5
-2,3,14.47,3.5
-2,5,26.01,4
-2,3,16.32,4
-2,4,21.33,4
-2,4,26.2,5
-2,3,20.25,5
-2,3,20.25,5
-2,2,16.28,6
-2,0,0,6
-2,1,8.31,6
-2,3,27.29,7
-2,2,18.98,7
-2,3,27.29,7
-2,2,21.41,8
-2,2,21.41,8
-2,2,21.41,8
-2,2,23.94,9
-2,2,23.94,9
-2,2,23.94,9
-2,2,26.38,10
-2,2,26.38,10
-2,2,26.38,10
-3,3,1.03,0.23
-3,2,0.69,0.23
-3,2,0.69,0.23
-3,5,2.35,0.5
-3,3,1.41,0.5
-3,5,2.35,0.5
-3,3,2.75,1
-3,4,3.66,1
-3,3,2.75,1
-3,4,5.57,1.5
-3,5,6.96,1.5
-3,6,8.33,1.5
-3,6,10.87,2
-3,5,9.09,2
-3,5,9.09,2
-3,6,13.63,2.5
-3,5,11.42,2.5
-3,7,15.79,2.5
-3,4,11.22,3
-3,4,11.22,3
-3,5,13.93,3
-3,4,13.06,3.5
-3,3,9.87,3.5
-3,4,13.06,3.5
-3,4,14.62,4
-3,4,14.62,4
-3,4,14.62,4
-3,2,9.43,5
-3,4,18.37,5
-3,3,13.98,5
-3,0,0,6
-3,0,0,6
-3,1,5.54,6
-3,1,6.45,7
-3,1,6.45,7
-3,4,24.32,7
-3,2,14.57,8
-3,5,33.02,8
-3,3,21.31,8
-3,4,30.63,9
-3,1,8.42,9
-3,4,30.63,9
-3,3,26.2,10
-3,3,26.2,10
-3,2,18.16,10
-1,3,1.03,0.23
-1,0,0,0.23
-1,3,1.03,0.23
-1,4,1.88,0.5
-1,6,2.82,0.5
-1,2,0.94,0.5
-1,4,3.66,1
-1,5,4.57,1
-1,5,4.57,1
-1,7,9.69,1.5
-1,5,6.96,1.5
-1,7,9.69,1.5
-1,6,10.87,2
-1,6,10.87,2
-1,4,7.29,2
-1,6,13.63,2.5
-1,6,13.63,2.5
-1,6,13.63,2.5
-1,5,13.93,3
-1,4,11.22,3
-1,6,16.57,3
-1,3,9.87,3.5
-1,7,22.1,3.5
-1,3,9.87,3.5
-1,4,14.62,4
-1,6,21.37,4
-1,6,21.37,4
-1,3,13.98,5
-1,5,22.54,5
-1,8,33.58,5
-1,0,0,6
-1,0,0,6
-1,0,0,6
-1,6,34.14,7
-1,5,29.47,7
-1,3,18.73,7
-1,5,33.02,8
-1,1,7.41,8
-1,5,33.02,8
-1,4,30.63,9
-1,3,23.94,9
-1,3,23.94,9
-1,2,18.16,10
-1,5,39.35,10
-1,2,18.16,10
-4,2,0.69,0.23
-4,4,1.37,0.23
-4,2,0.69,0.23
-4,6,4.12,0.5
-4,4,2.75,0.5
-4,2,1.37,0.5
-4,2,2.79,1
-4,3,4.19,1
-4,3,4.19,1
-4,4,8.28,1.5
-4,2,4.16,1.5
-4,3,6.23,1.5
-4,3,8.19,2
-4,2,5.48,2
-4,3,8.19,2
-4,5,16.8,2.5
-4,4,13.58,2.5
-4,3,10.27,2.5
-4,3,12.45,3
-4,4,16.4,3
-4,3,12.45,3
-4,5,23.27,3.5
-4,3,14.47,3.5
-4,3,14.47,3.5
-4,3,16.32,4
-4,4,21.33,4
-4,3,16.32,4
-4,5,31.59,5
-4,2,13.82,5
-4,3,20.25,5
-4,0,0,6
-4,4,30.28,6
-4,1,8.31,6
-4,2,18.98,7
-4,3,27.29,7
-4,2,18.98,7
-4,2,21.41,8
-4,2,21.41,8
-4,2,21.41,8
-4,2,23.94,9
-4,2,23.94,9
-4,2,23.94,9
-4,3,36.65,10
-4,2,26.38,10
-4,1,13.93,10
