--- T1/Data/jane_11-7_T1.csv
@@ -1,181 +1,0 @@
-Direction, Characters Read, Eccentricity (degrees), Letter Height (degrees)
-2,0,0,0.23
-2,1,0.34,0.23
-2,3,1.03,0.23
-3,1,0.34,0.23
-3,2,0.69,0.23
-3,0,0,0.23
-1,0,0,0.23
-1,8,2.75,0.23
-1,5,1.72,0.23
-1,0,0,0.23
-4,1,0.34,0.23
-4,4,1.37,0.23
-4,3,1.03,0.23
-2,3,2.06,0.5
-2,5,3.43,0.5
-2,8,5.48,0.5
-3,5,2.35,0.5
-3,9,4.22,0.5
-3,9,4.22,0.5
-1,10,4.69,0.5
-1,9,4.22,0.5
-4,6,4.12,0.5
-4,4,2.75,0.5
-4,5,3.43,0.5
-4,8,11.05,0.5
-2,6,8.33,1
-2,7,9.69,1
-2,4,5.57,1
-3,2,1.83,1
-3,7,6.39,1
-3,5,4.57,1
-3,5,6.96,1
-1,8,7.29,1
-1,11,9.98,1
-1,7,6.39,1
-4,6,8.33,1
-4,4,5.57,1
-2,6,12.32,1.5
-2,5,10.31,1.5
-2,1,2.08,1.5
-3,4,5.57,1.5
-3,8,11.05,1.5
-1,4,5.57,1.5
-1,8,11.05,1.5
-1,8,11.05,1.5
-4,0,0,1.5
-4,0,0,1.5
-4,3,6.23,1.5
-4,4,10.87,1.5
-2,5,13.5,2
-2,1,2.75,2
-2,0,0,2
-2,3,10.27,2
-3,5,9.09,2
-3,5,9.09,2
-3,8,14.36,2
-1,9,16.07,2
-1,8,14.36,2
-1,9,16.07,2
-4,5,13.5,2
-4,2,5.48,2
-2,6,19.92,2.5
-2,5,16.8,2.5
-3,3,6.91,2.5
-3,8,17.91,2.5
-3,4,9.18,2.5
-1,8,17.91,2.5
-1,6,13.63,2.5
-1,7,15.79,2.5
-1,3,8.46,2.5
-4,4,13.58,2.5
-4,3,10.27,2.5
-4,3,10.27,2.5
-2,4,16.4,3
-2,5,20.2,3
-2,4,16.4,3
-2,3,14.47,3
-3,4,11.22,3
-3,6,16.57,3
-3,4,11.22,3
-1,8,21.64,3
-1,3,8.46,3
-4,3,12.45,3
-4,3,12.45,3
-4,2,8.37,3
-2,0,0,3.5
-2,2,9.76,3.5
-2,6,30.35,3.5
-3,6,19.19,3.5
-3,6,19.19,3.5
-3,4,13.06,3.5
-1,4,13.06,3.5
-1,10,30.11,3.5
-1,5,16.17,3.5
-4,2,9.76,3.5
-4,4,18.98,3.5
-4,5,23.27,3.5
-2,4,21.33,4
-2,1,5.57,4
-3,2,7.43,4
-3,6,21.37,4
-3,4,14.62,4
-1,6,21.37,4
-1,6,21.37,4
-1,6,21.37,4
-4,3,16.32,4
-4,3,16.32,4
-4,2,11.05,4
-4,2,13.82,4
-2,4,26.2,5
-2,2,13.82,5
-2,4,26.2,5
-2,3,23.65,5
-3,4,18.37,5
-3,5,22.54,5
-3,4,18.37,5
-1,5,22.54,5
-1,4,18.37,5
-1,4,18.37,5
-4,0,0,5
-4,0,0,5
-2,1,8.31,6
-2,2,16.28,6
-3,3,16.22,6
-3,5,25.87,6
-3,5,25.87,6
-1,3,16.22,6
-1,7,34.18,6
-1,4,21.21,6
-4,4,30.28,6
-4,2,16.28,6
-4,2,16.28,6
-4,3,27.29,6
-2,3,27.29,7
-2,3,27.29,7
-2,2,18.98,7
-3,3,18.73,7
-3,5,29.47,7
-3,4,24.32,7
-1,3,18.73,7
-1,3,18.73,7
-1,6,34.14,7
-1,3,21.31,7
-4,2,18.98,7
-4,2,18.98,7
-2,2,21.41,8
-2,2,21.41,8
-2,2,21.41,8
-3,4,27.47,8
-3,3,21.31,8
-3,2,14.57,8
-1,4,27.47,8
-1,6,37.95,8
-1,5,36.5,8
-4,2,21.41,8
-4,2,21.41,8
-4,2,21.41,8
-2,2,23.94,9
-2,2,23.94,9
-2,2,23.94,9
-2,2,26.38,9
-3,5,36.5,9
-3,5,36.5,9
-3,5,36.5,9
-1,0,0,9
-1,4,30.63,9
-4,2,23.94,9
-4,2,23.94,9
-4,2,23.94,9
-2,2,26.38,10
-2,2,26.38,10
-3,3,26.2,10
-3,3,26.2,10
-3,5,39.35,10
-1,3,26.2,10
-1,2,18.16,10
-1,5,39.35,10
-4,2,26.38,10
-4,3,36.65,10
-4,3,36.65,10
