--- T1/Data/sara_11-7.csv
@@ -1,181 +1,0 @@
-Direction, Characters Read, Eccentricity (degrees), Letter Height (degrees)
-2,2,0.69,0.23
-2,0,0,0.23
-2,4,1.37,0.23
-2,1,0.69,0.5
-2,1,0.69,0.5
-2,0,0,0.5
-2,10,13.71,1
-2,7,9.69,1
-2,2,2.79,1
-2,0,0,1.5
-2,1,2.08,1.5
-2,0,0,1.5
-2,7,18.57,2
-2,4,10.87,2
-2,5,13.5,2
-2,4,13.58,2.5
-2,4,13.58,2.5
-2,0,0,2.5
-2,6,23.83,3
-2,4,16.4,3
-2,2,8.37,3
-2,3,14.47,3.5
-2,0,0,3.5
-2,6,27.29,3.5
-2,6,30.35,4
-2,2,11.05,4
-2,1,5.57,4
-2,0,0,5
-2,3,20.25,5
-2,0,0,5
-2,2,16.28,6
-2,3,23.65,6
-2,0,0,6
-2,3,27.29,7
-2,3,27.29,7
-2,0,0,7
-2,0,0,8
-2,2,21.41,8
-2,2,21.41,8
-2,1,12.52,9
-2,2,23.94,9
-2,2,23.94,9
-2,2,26.38,10
-2,2,26.38,10
-2,2,26.38,10
-3,4,1.37,0.23
-3,0,0,0.23
-3,2,0.69,0.23
-3,1,0.47,0.5
-3,0,0,0.5
-3,3,1.41,0.5
-3,2,1.83,1
-3,3,2.75,1
-3,1,0.92,1
-3,9,12.39,1.5
-3,1,1.4,1.5
-3,2,2.79,1.5
-3,1,1.83,2
-3,0,0,2
-3,1,1.83,2
-3,0,0,2.5
-3,0,0,2.5
-3,1,2.31,2.5
-3,7,19.15,3
-3,7,19.15,3
-3,6,16.57,3
-3,5,16.17,3.5
-3,2,6.62,3.5
-3,0,0,3.5
-3,5,18.06,4
-3,2,7.43,4
-3,1,3.73,4
-3,6,26.47,5
-3,9,36.76,5
-3,0,0,5
-3,4,21.21,6
-3,0,0,6
-3,2,10.98,6
-3,4,24.32,7
-3,3,18.73,7
-3,1,6.45,7
-3,5,33.02,8
-3,3,21.31,8
-3,0,0,8
-3,6,41.61,9
-3,4,30.63,9
-3,2,16.49,9
-3,3,26.2,10
-3,1,9.31,10
-3,5,39.35,10
-1,10,3.43,0.23
-1,1,0.34,0.23
-1,1,0.34,0.23
-1,2,0.94,0.5
-1,0,0,0.5
-1,3,1.41,0.5
-1,0,0,1
-1,2,1.83,1
-1,0,0,1
-1,1,1.4,1.5
-1,1,1.4,1.5
-1,0,0,1.5
-1,1,1.83,2
-1,6,10.87,2
-1,0,0,2
-1,6,13.63,2.5
-1,3,6.91,2.5
-1,1,2.31,2.5
-1,0,0,3
-1,3,8.46,3
-1,5,13.93,3
-1,6,19.19,3.5
-1,6,19.19,3.5
-1,3,9.87,3.5
-1,8,27.55,4
-1,0,0,4
-1,0,0,4
-1,1,4.74,5
-1,0,0,5
-1,3,13.98,5
-1,3,16.22,6
-1,3,16.22,6
-1,2,10.98,6
-1,3,18.73,7
-1,2,12.73,7
-1,1,6.45,7
-1,6,37.95,8
-1,3,21.31,8
-1,2,14.57,8
-1,4,30.63,9
-1,3,23.94,9
-1,4,30.63,9
-1,4,33.26,10
-1,0,0,10
-1,5,39.35,10
-4,3,1.03,0.23
-4,2,0.69,0.23
-4,1,0.34,0.23
-4,2,1.37,0.5
-4,1,0.69,0.5
-4,1,0.69,0.5
-4,3,4.19,1
-4,0,0,1
-4,0,0,1
-4,3,6.23,1.5
-4,0,0,1.5
-4,1,2.08,1.5
-4,4,10.87,2
-4,3,8.19,2
-4,2,5.48,2
-4,2,6.89,2.5
-4,1,3.46,2.5
-4,7,22.92,2.5
-4,4,16.4,3
-4,3,12.45,3
-4,3,12.45,3
-4,4,18.98,3.5
-4,3,14.47,3.5
-4,2,9.76,3.5
-4,3,16.32,4
-4,5,26.01,4
-4,6,30.35,4
-4,3,20.25,5
-4,4,26.2,5
-4,4,26.2,5
-4,2,16.28,6
-4,0,0,6
-4,4,30.28,6
-4,3,27.29,7
-4,3,27.29,7
-4,1,9.76,7
-4,2,21.41,8
-4,2,21.41,8
-4,1,11.09,8
-4,2,23.94,9
-4,0,0,9
-4,2,23.94,9
-4,0,0,10
-4,2,26.38,10
-4,2,26.38,10
