--- T1/Data/thida_11-7.csv
@@ -1,181 +1,0 @@
-Direction, Characters Read, Eccentricity (degrees), Letter Height (degrees)
-2,1,0.34,0.23
-2,0,0,0.23
-2,3,1.03,0.23
-2,0,0,0.5
-2,1,0.69,0.5
-2,2,1.37,0.5
-2,5,6.96,1
-2,2,2.79,1
-2,5,6.96,1
-2,6,12.32,1.5
-2,4,8.28,1.5
-2,7,14.29,1.5
-2,4,10.87,2
-2,5,13.5,2
-2,0,0,2
-2,5,16.8,2.5
-2,5,16.8,2.5
-2,5,16.8,2.5
-2,4,16.4,3
-2,3,12.45,3
-2,4,16.4,3
-2,3,14.47,3.5
-2,4,18.98,3.5
-2,5,23.27,3.5
-2,3,16.32,4
-2,0,0,4
-2,0,0,4
-2,4,26.2,5
-2,4,26.2,5
-2,3,20.25,5
-2,3,23.65,6
-2,3,23.65,6
-2,3,23.65,6
-2,3,27.29,7
-2,2,18.98,7
-2,2,18.98,7
-2,2,21.41,8
-2,2,21.41,8
-2,2,21.41,8
-2,2,23.94,9
-2,2,23.94,9
-2,2,23.94,9
-2,2,26.38,10
-2,2,26.38,10
-2,2,26.38,10
-3,6,2.06,0.23
-3,1,0.34,0.23
-3,3,1.03,0.23
-3,1,0.47,0.5
-3,1,0.47,0.5
-3,5,2.35,0.5
-3,4,3.66,1
-3,2,1.83,1
-3,2,1.83,1
-3,6,8.33,1.5
-3,4,5.57,1.5
-3,5,6.96,1.5
-3,7,12.63,2
-3,0,0,2
-3,3,5.48,2
-3,6,13.63,2.5
-3,7,15.79,2.5
-3,0,0,2.5
-3,4,11.22,3
-3,4,11.22,3
-3,4,11.22,3
-3,3,9.87,3.5
-3,4,13.06,3.5
-3,5,16.17,3.5
-3,5,18.06,4
-3,5,18.06,4
-3,7,24.53,4
-3,5,22.54,5
-3,3,13.98,5
-3,4,18.37,5
-3,8,37.81,6
-3,3,16.22,6
-3,5,25.87,6
-3,6,34.14,7
-3,3,18.73,7
-3,5,29.47,7
-3,4,27.47,8
-3,7,42.3,8
-3,3,21.31,8
-3,6,41.61,9
-3,2,16.49,9
-3,5,36.5,9
-3,4,33.26,10
-3,3,26.2,10
-3,5,39.35,10
-1,6,2.06,0.23
-1,0,0,0.23
-1,7,2.41,0.23
-1,1,0.47,0.5
-1,2,0.94,0.5
-1,4,1.88,0.5
-1,3,2.75,1
-1,5,4.57,1
-1,4,3.66,1
-1,4,5.57,1.5
-1,3,4.19,1.5
-1,6,8.33,1.5
-1,5,9.09,2
-1,7,12.63,2
-1,5,9.09,2
-1,3,6.91,2.5
-1,4,9.18,2.5
-1,5,11.42,2.5
-1,5,13.93,3
-1,4,11.22,3
-1,4,11.22,3
-1,6,19.19,3.5
-1,4,13.06,3.5
-1,3,9.87,3.5
-1,0,0,4
-1,4,14.62,4
-1,7,24.53,4
-1,4,18.37,5
-1,6,26.47,5
-1,4,18.37,5
-1,3,16.22,6
-1,5,25.87,6
-1,4,21.21,6
-1,1,6.45,7
-1,4,24.32,7
-1,6,34.14,7
-1,6,37.95,8
-1,4,27.47,8
-1,6,37.95,8
-1,5,36.5,9
-1,2,16.49,9
-1,3,23.94,9
-1,4,33.26,10
-1,5,39.35,10
-1,5,39.35,10
-4,2,0.69,0.23
-4,2,0.69,0.23
-4,1,0.34,0.23
-4,2,1.37,0.5
-4,2,1.37,0.5
-4,3,2.06,0.5
-4,2,2.79,1
-4,2,2.79,1
-4,2,2.79,1
-4,5,10.31,1.5
-4,0,0,1.5
-4,1,2.08,1.5
-4,8,21.01,2
-4,2,5.48,2
-4,3,8.19,2
-4,3,10.27,2.5
-4,3,10.27,2.5
-4,3,10.27,2.5
-4,3,12.45,3
-4,2,8.37,3
-4,4,16.4,3
-4,2,9.76,3.5
-4,0,0,3.5
-4,4,18.98,3.5
-4,5,26.01,4
-4,2,11.05,4
-4,3,16.32,4
-4,2,13.82,5
-4,5,31.59,5
-4,3,20.25,5
-4,2,16.28,6
-4,3,23.65,6
-4,3,23.65,6
-4,2,18.98,7
-4,3,27.29,7
-4,3,27.29,7
-4,2,21.41,8
-4,1,11.09,8
-4,2,21.41,8
-4,1,12.52,9
-4,1,12.52,9
-4,2,23.94,9
-4,3,36.65,10
-4,2,26.38,10
-4,2,26.38,10
